requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
Flask-SQLAlchemy==3.1.1
ipykernel==6.29.5
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
wcwidth==0.2.13
wheel @ file:///usr/local/Cellar/python%403.12/3.12.6/libexec/wheel-0.44.0-py3-none-any.whl#sha256=d6ad895f80a568fbc066fd6c52c68c5ef40632c4d764d42af13e28e945f1920b
